Split ChatComposer into lane picker + dictation hook (377 → 183)
- ChatComposer.tsx                  shell — composer state, textarea
                                    autosize, send/abort buttons, edit-
                                    mode banner
- chat-composer/lanes.ts            LANE_OPTIONS constant + LaneOption
                                    type
- chat-composer/LanePicker.tsx      lane chip + dropdown menu + Flow
                                    Mode link. Owns its own open state
                                    and click-outside dismissal so the
                                    composer parent doesn't need to.
- chat-composer/useDictation.ts     Web Speech API hook — returns
                                    isDictating + speechSupported +
                                    toggle. Caller hooks a transcript
                                    callback for the textarea.

All files under 190 lines.
